MRP-57 xattr: a few fixes for the XATTR cache
This patch separates ACL and XATTR caches, so that
when updating an ACL only LOOKUP lock is needed and
when updating another XATTR only XATTR lock is needed.

This patch also reverts XATTR cache support for setxattr
because a client performing REINT under even PR lock
can deadlock if an active server operation (like unlink)
attempts to cancel all locks, and setxattr has to wait
for it (MDC max-in-flight is 1).

In some cases, MDS updates extended attributes
on its own rather than on a client request.
MDS_INODELOCK_XATTR should be taken for this purpose.

This patch also sets up HABEO_REFERO flag for the
getxattr intent alike getxattr REINT operation.
Without it, Lustre can crash on "non-existent object"
assertion.

This patch disables the r/o cache if the data is
unreasonably large (larger than maximum single EA
size).

Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/605
Tested-by: Jenkins
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>

+++ lustre/tests/sanityn.sh
@@ -1953,19 +1953,29 @@ test_51() {
 run_test 51 "alloc_rr should be allocate on correct order"
 
 test_52() {
-        local p="$TMP/sanityN-$TESTNAME.parameters"
-        save_lustre_params $HOSTNAME "llite.*.xattr_cache" > $p
-        lctl set_param llite.*.xattr_cache 1 || { skip "xattr cache is not supported"; return 0; }
+	local tlink1
+	local tlink2
+	local p="$TMP/sanityN-$TESTNAME.parameters"
+	save_lustre_params $HOSTNAME "llite.*.xattr_cache" > $p
+	lctl set_param llite.*.xattr_cache 1 || { skip "xattr cache is not supported"; return 0; }
 
-        touch $DIR1/$tfile
-        setfattr -n user.attr1 -v value1 $DIR1/$tfile || error 1
-        getfattr -n user.attr1 $DIR2/$tfile | grep value1 || error 2
-        setfattr -n user.attr1 -v value2 $DIR2/$tfile || error 3
-        getfattr -n user.attr1 $DIR1/$tfile | grep value2 || error 4
-        rm -f $DIR2/$tfile
+	touch $DIR1/$tfile
+	setfattr -n user.attr1 -v value1 $DIR1/$tfile || error 1
+	getfattr -n user.attr1 $DIR2/$tfile | grep value1 || error 2
+	setfattr -n user.attr1 -v value2 $DIR2/$tfile || error 3
+	getfattr -n user.attr1 $DIR1/$tfile | grep value2 || error 4
 
-        restore_lustre_params < $p
-        rm -f $p
+	# check that trusted.link is consistent
+	tlink1=$(getfattr -n trusted.link $DIR1/$tfile | md5sum)
+	ln $DIR2/$tfile $DIR2/$tfile-2 || error "failed to link"
+	tlink2=$(getfattr -n trusted.link $DIR1/$tfile | md5sum)
+	echo "$tlink1 $tlink2"
+	[ "$tlink1" = "$tlink2" ] && error "trusted.link should have changed!"
+	rm -f $DIR2/$tfile
+	rm -f $DIR2/$tfile-2
+
+	restore_lustre_params < $p
+	rm -f $p
 }
 run_test 52 "getxattr/setxattr cache should be consistent between nodes"
